requirements-dev.txt
-r requirements.txt
pytest-django==3.1.2
codecov==2.0.9
flake8==3.4.1
Fabric3==1.13.1.post1
requests-mock==1.3.0
six==1.11.0